% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cowboy_glm.R
\name{cowboy_glm}
\alias{cowboy_glm}
\title{Run glm with Lasso prescreening built into boostrapped confidence intervals}
\usage{
cowboy_glm(
  data,
  clusterid = "id",
  Ws,
  forcedW = NULL,
  pair = NULL,
  family = "gaussian",
  B = 200,
  confint.level = 0.95,
  n.cores = 1
)
}
\arguments{
\item{Ws}{data frame that includes candidate adjustment covariates to screen}

\item{family}{GLM model family (gaussian, binomial, poisson, or negative binomial). Use "neg.binom" for Negative binomial.}

\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{pval}{The p-value threshold: any variables with a p-value from the lielihood ratio test below this threshold will be returned. Defaults to 0.2}

\item{print}{Logical for whether to print function output, defaults to TRUE.}
}
\value{
Function returns the list of variable names with a likelihood ratio test p-value <0.2 (unless a custom p-value is specified).
}
\description{
Run glm with Lasso prescreening built into boostrapped confidence intervals
}
\examples{

}
