% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_matrix.R
\name{design_matrix}
\alias{design_matrix}
\title{Convert a vector, matrix or data frame with factors into a design matrix}
\usage{
design_matrix(W)
}
\arguments{
\item{W}{A vector, matrix, or data.frame that includes numeric or factor variables.}
}
\value{
A design matrix version of \code{W} where factor variables have been converted into columns of indicator variables with the first level excluded.
}
\description{
An internal, convenience function that automatically transforms a vector, matrix, or data.frame with factors into a design matrix with indicator variables and an ommitted category. This is handy for using the SuperLearner and tmle packages
}
\details{
The \code{design_matrix} function is used by \code{agecurveAb} and \code{tmleAb} as a data processing tool for functions that do not readily accommodate factor variables in a data.frame object. In particular, the \code{SuperLearner()} and \code{tmle()} functions typically have difficulty with factor variables and so this function transforms data before calling those functions.
}
\keyword{internal}
