% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_ITT.unadj.R
\name{ITT.unadj}
\alias{ITT.unadj}
\title{washb_ITT.unadj}
\usage{
ITT.unadj(Y, tr, strat, contrast, binomial = FALSE, measure = "RR")
}
\arguments{
\item{Y}{binary outcome variable (here: diar7d)}

\item{tr}{binary treatment group variable, comparison group first}

\item{strat}{stratification variable (here: block)}

\item{contrast}{vector of length 2 that includes the tr groups to contrast}

\item{binomial}{logicial. If TRUE, then the M-H estimator is used for binary outcomes}

\item{measure}{measure of effect for the M-H estimator. RR = prev ratio, RD = prev difference. ignored unless binomial==TRUE.}
}
\value{
to be written
}
\description{
Unadjusted estimates for continuous (paired t-test) or binomial (mantel-haenszel) outcomes
}
\details{
unadjusted ITT estimates of differences between arms, calculated with either a paired T-test
(continuous outcomes) or a Mantel-Haenszel pooled estimator (binary outcomes).
The arguments Y,tr,strat, below need to be from the same dataset
}
\examples{
 to be written

}

