% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadUnblindedData.R
\name{loadUnblindedData}
\alias{loadUnblindedData}
\title{Internal function to load unblinded .sysdata into the vignette}
\usage{
loadUnblindedData(dataname)
}
\description{
Internal function to load unblinded .sysdata into the vignette
}

