% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_paired.ttest.R
\name{paired.ttest}
\alias{paired.ttest}
\title{washb_paired.ttest}
\usage{
paired.ttest(Y, tr, strat, contrast)
}
\arguments{
\item{Y}{quantitative outcome variable (e.g. LAZ)}

\item{tr}{binary treatment group variable, comparison group first}

\item{strat}{stratification variable (here: block)}

\item{contrast}{vector of length 2 that includes the tr groups to contrast}
}
\value{
to be written
}
\description{
Estimate paired t-tests for differences in means at the randomization block level
}
\details{
Function to call the paired t-test for two different arms of the study
Estimate the paired t-test for differences in means paired within randomization blocks.
The arguments Y,tr,strat, below need to be from the same dataset:
}
\examples{
 to be written

}

