% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwichSE.R
\name{sandwichSE}
\alias{sandwichSE}
\title{sandwichSE}
\usage{
sandwichSE(dat, fm, cluster)
}
\arguments{
\item{dat}{Data used to fit the model}

\item{fm}{Model fit (object)}

\item{cluster}{Vector of cluster ID's}
}
<<<<<<< HEAD
\value{
to be written
}
=======
>>>>>>> ad784f1a15d88f88351ea24f45f9ef11ca46bb47
\description{
sandwichSE
}
\examples{
<<<<<<< HEAD
 to be written
=======
>>>>>>> ad784f1a15d88f88351ea24f45f9ef11ca46bb47

}

