% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_bang_tr.R
\docType{data}
\name{washb_bang_tr}
\alias{washb_bang_tr}
\title{dataset: Bangladesh treatment assignments}
\format{A data frame with 5551 observations and 8 variables.
\describe{
 \item{block}{Randomization block ID}
 \item{clusterid}{Cluster ID)}
 \item{tr}{Treatment assignment}
 }}
\usage{
data(washb_bang_tr)
}
\description{
Bangladesh treatment assignments dataset
}
\keyword{datasets}
