% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_bangladesh_enrol.R
\docType{data}
\name{washb_bangladesh_enrol}
\alias{washb_bangladesh_enrol}
\title{dataset: Bangladesh enrollment analysis}
\format{
A data frame with 5,551 observations and 73 variables.
\describe{
 \item{\code{dataid}}{Compound ID}
 \item{\code{clusterid}}{Cluster ID)}
 \item{\code{hhid}}{Household ID}
 \item{\code{block}}{Randomization block ID}
 \item{\code{tr}}{Randomized treatment assignment}
 \item{\code{union}}{Administrative union (q4008)}
 \item{\code{fracode}}{FRA code (q4001)}
 \item{\code{svydate}}{Survey date}
 \item{\code{Nhh}}{Number of persons in household (q012)}
 \item{\code{Nlt18}}{N indivs in HH <=18y}
 \item{\code{Ncomp}}{Total indivs in compound}
 \item{\code{momage}}{Mother's age, years}
 \item{\code{momheight}}{Maternal height (median)}
 \item{\code{momedu}}{Mother's education (category)}
 \item{\code{momeduy}}{Mother, years of education (q009)}
 \item{\code{dadeduy}}{Father, years of education (q010)}
 \item{\code{dadagri}}{Father works in agriculture (q011)}
 \item{\code{landacre}}{Acres of land owned (q4110)}
 \item{\code{hfias}}{HFIAs food insecurity score (0-27)}
 \item{\code{hfiacat}}{HFIAS food insecurity category}
 \item{\code{tubewell}}{Primary water source: shallow tubewell (q1010)}
 \item{\code{watmin}}{Dist (mins) to primary water source}
 \item{\code{storewat}}{Store drinking water (q1003)}
 \item{\code{treatwat}}{Reported treating water today or yesterday (q1008)}
 \item{\code{odmen}}{Daily OD, adult men (q801a)}
 \item{\code{odwom}}{Daily OD, adult women (q801b)}
 \item{\code{odch815}}{Daily OD, children 8-15 (q801e)}
 \item{\code{odch38}}{Daily OD, children 3-8 (q801d)}
 \item{\code{odchu3}}{Daily OD, children <3 (q801c)}
 \item{\code{latown}}{Own their latrine (not shared) (q816)}
 \item{\code{latslab}}{Latrine has concrete slab (q809_7)}
 \item{\code{latseal}}{Latrine has functional water seal (q809_9a)}
 \item{\code{latfeces}}{No visible feces on slab/floor of latrine}
 \item{\code{potty}}{Has a potty for child defecation}
 \item{\code{humfeces}}{Human feces observed in house/compound (q4201,4203,4205)}
 \item{\code{humfecesch}}{Human feces obs in loc where child spends most time (q4203)}
 \item{\code{hwlat}}{Has handwashing loc w/in 6 steps of latrine}
 \item{\code{hwlatwat}}{Has handwashing loc w/in 6 steps of latrine w/ water}
 \item{\code{hwlatsoap}}{Has handwashing loc w/in 6 steps of latrine w/ soap}
 \item{\code{hwkit}}{Has handwashing loc w/in 6 steps of kitchen}
 \item{\code{hwkitwat}}{Has handwashing loc w/in 6 steps of kitchen w/ water}
 \item{\code{hwkitsoap}}{Has handwashing loc w/in 6 steps of kitchen w/ soap}
 \item{\code{hwsw}}{Prim handwashing loc has water (q704_1)}
 \item{\code{hwss}}{Prim handwashing loc has soap (q704_2-6)}
 \item{\code{hwsws}}{Prim handwashing loc has water+soap (q704_1-6)}
 \item{\code{roof}}{Improved roof material (tin, cement)}
 \item{\code{walls}}{Improved wall material (wood, brick, tin)}
 \item{\code{floor}}{Improved floor (wood, concrete)}
 \item{\code{cement}}{Has cement floor (q4103)}
 \item{\code{elec}}{Household has electricity}
 \item{\code{asset_radio}}{households have radio}
 \item{\code{asset_tvbw}}{households have B/W television}
 \item{\code{asset_tvcol}}{households have color television}
 \item{\code{asset_refrig}}{households have Refrigerator}
 \item{\code{asset_bike}}{households have bicycle}
 \item{\code{asset_moto}}{households have motorcycle}
 \item{\code{asset_sewmach}}{households have sewing machine}
 \item{\code{asset_phone}}{number of land phone}
 \item{\code{asset_tv}}{Has bw or color TV}
 \item{\code{asset_wardrobe}}{Has >=1 wardrobe}
 \item{\code{asset_table}}{Has >=1 table}
 \item{\code{asset_chair}}{Has >=1 chair}
 \item{\code{asset_clock}}{Has >=1 clock}
 \item{\code{asset_khat}}{Has >=1 khat}
 \item{\code{asset_chouki}}{Has >=1 chouki}
 \item{\code{asset_mobile}}{Has >=1 mobile}
 \item{\code{n_asset_wardrobe}}{number of wardrobes}
 \item{\code{n_asset_table}}{number of Tables}
 \item{\code{n_asset_chair}}{number of Chair}
 \item{\code{n_asset_clock}}{number of clock}
 \item{\code{n_asset_khat}}{number of Khat}
 \item{\code{n_asset_chouki}}{number of chouki}
 \item{\code{n_asset_mobile}}{number of mobile phone}
 }
}
\usage{
data(washb_bangladesh_enrol)
}
\description{
Bangladesh enrollment analysis dataset
}
\keyword{datasets}
