% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_bangladesh_track_compound.R
\docType{data}
\name{washb_bangladesh_track_compound}
\alias{washb_bangladesh_track_compound}
\title{dataset: Bangladesh track compound analysis}
\format{A data frame with 5551 observations and 8 variables.
\describe{
 \item{\code{dataid}}{Compound ID}
 \item{\code{clusterid}}{Cluster ID)}
 \item{\code{block}}{Randomization block ID}
 \item{\code{tr}}{Treatment assignment}
 \item{\code{miss1}}{Miss 1}
 \item{\code{miss1reason}}{Miss 1 Reason}
 \item{\code{miss2}}{Miss 2}
 \item{\code{miss2reason}}{Miss 2 Reason}
 }}
\usage{
data(washb_bangladesh_track_compound)
}
\description{
Bangladesh track compound analysis
}
\keyword{datasets}
