% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_bd_anthro.R
\docType{data}
\name{washb_bd_anthro}
\alias{washb_bd_anthro}
\title{Anthropometry data from Wash Benefits Bangladesh Children}
\format{A data frame with 9342 rows and 46 variables:
\describe{
  \item{dataid}{Unique observation ID}
  \item{childid}{Child ID}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}


  ...
}}
\usage{
washb_bd_anthro
}
\description{
A dataset containing circumference, length, weight, and age data for
Bangladesh Wash Benefits children.
}
\keyword{datasets}

