% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_bd_comp.R
\docType{data}
\name{washb_bd_comp}
\alias{washb_bd_comp}
\title{Dataset tracking drop outs of enrolled Wash Benefits Bangladesh compounds}
\format{A data frame with 5551 rows and 8 variables:
\describe{
  \item{dataid}{dataid}
  \item{clusterid}{Cluster ID}
  \item{block}{Randomization block ID}
  \item{tr}{Randomized treatment assignment}
  \item{miss1}{Not measured at year 1}
  \item{miss1reason}{Reason for no measurement in year 1}
  \item{miss2}{Not measured at year 2}
  \item{miss2reason}{Reason for no measurement in year 2}


  ...
}}
\usage{
washb_bd_comp
}
\description{
WASH Benefits Bangladesh tracking file, compounds
}
\keyword{datasets}

