% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_bd_enrol.R
\docType{data}
\name{washb_bd_enrol}
\alias{washb_bd_enrol}
\title{Enrollment characteristics for Wash Benefits Bangladesh Households}
\format{A data frame with 5551 rows and 73 variables:
\describe{
  \item{dataid}{dataid}
  \item{clusterid}{Cluster ID}
  \item{hhid}{Household ID}
  \item{block}{Randomization block ID}
  \item{tr}{Randomized treatment assignment}
  \item{union}{Administrative union (q4008)}
  \item{fracode}{FRA code (q4001)}
  \item{svydate}{Survey date}
  \item{Nhh}{Number of persons in household (q012)}
  \item{Nlt18}{N indivs in HH <=18y}
  \item{Ncomp}{Total indivs in compound}
  \item{momage}{Mother's age, years}
  \item{momheight}{Maternal height (median)}
  \item{momedu}{Mother's education (category)}
  \item{momeduy}{Mother, years of education (q009)}
  \item{dadeduy}{Father, years of education (q010)}
  \item{dadagri}{Father works in agriculture (q011)}
  \item{landacre}{Acres of land owned (q4110)}
  \item{hfias}{HFIAs food insecurity score (0-27)}
  \item{hfiacat}{HFIAS food insecurity category}
  \item{tubewell}{Primary water source: shallow tubewell (q1010)}
  \item{watmin}{Dist (mins) to primary water source}
  \item{storewat}{Store drinking water (q1003)}
  \item{treatwat}{Reported treating water today or yesterday (q1008)}
  \item{odmen}{Daily OD, adult men (q801a)}
  \item{odwom}{Daily OD, adult women (q801b)}
  \item{odch815}{Daily OD, children 8-15 (q801e)}
  \item{odch38}{Daily OD, children 3-8 (q801d)}
  \item{odchu3}{Daily OD, children <3 (q801c)}
  \item{latown}{Own their latrine (not shared) (q816)}
  \item{latslab}{Latrine has concrete slab (q809_7)}
  \item{latseal}{Latrine has functional water seal (q809_9a)}
  \item{latfeces}{No visible feces on slab/floor of latrine}
  \item{potty}{Has a potty for child defecation}
  \item{humfeces}{Human feces observed in house/compound (q4201,4203,4205)}
  \item{humfecesch}{Human feces obs in loc where child spends most time (q4203)}
  \item{hwlat}{Has handwashing loc w/in 6 steps of latrine}
  \item{hwlatwat}{Has handwashing loc w/in 6 steps of latrine w/ water}
  \item{hwlatsoap}{Has handwashing loc w/in 6 steps of latrine w/ soap}
  \item{hwkit}{Has handwashing loc w/in 6 steps of kitchen}
  \item{hwkitwat}{Has handwashing loc w/in 6 steps of kitchen w/ water}
  \item{hwkitsoap}{Has handwashing loc w/in 6 steps of kitchen w/ soap}
  \item{hwsw}{Prim handwashing loc has water (q704_1)}
  \item{hwss}{Prim handwashing loc has soap (q704_2-6)}
  \item{hwsws}{Prim handwashing loc has water+soap (q704_1-6)}
  \item{roof}{Improved roof material (tin, cement)}
  \item{walls}{Improved wall material (wood, brick, tin)}
  \item{floor}{Improved floor (wood, concrete)}
  \item{cement}{Has cement floor (q4103)}
  \item{elec}{Household has electricity}
  \item{asset_radio}{households have radio}
  \item{asset_tvbw}{households have B/W television}
  \item{asset_tvcol}{households have color television}
  \item{asset_refrig}{households have Refrigerator}
  \item{asset_bike}{households have bicycle}
  \item{asset_moto}{households have motorcycle}
  \item{asset_sewmach}{households have sewing machine}
  \item{asset_phone}{number of land phone}
  \item{asset_tv}{Has bw or color TV}
  \item{asset_wardrobe}{Has >=1 wardrobe}
  \item{asset_table}{Has >=1 table}
  \item{asset_chair}{Has >=1 chair}
  \item{asset_clock}{Has >=1 clock}
  \item{asset_khat}{Has >=1 khat}
  \item{asset_chouki}{Has >=1 chouki}
  \item{asset_mobile}{Has >=1 mobile}
  \item{n_asset_wardrobe}{number of waredrobe}
  \item{n_asset_table}{number of Tables}
  \item{n_asset_chair}{number of Chair}
  \item{n_asset_clock}{number of clock}
  \item{n_asset_khat}{number of Khat}
  \item{n_asset_chouki}{number of chouki}
  \item{n_asset_mobile}{number of mobile phone}


  ...
}}
\source{
NA
}
\usage{
washb_bd_enrol
}
\description{
Bangladesh enrollment analysis dataset (HH obs), created by 2-bangladesh-dm-enro
}
\keyword{datasets}

