% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_bd_uptake.R
\docType{data}
\name{washb_bd_uptake}
\alias{washb_bd_uptake}
\title{Intervention Uptake data from Wash Benefits Bangladesh Households}
\format{A data frame with 14908 rows and 18 variables:
\describe{
  \item{dataid}{Compound ID}
  \item{clusterid}{Cluster ID}
  \item{block}{Randomization block ID}
  \item{tr}{Treatment assignment}
  \item{svy}{Survey round (0,1,2)}
  \item{svydate}{Survey date}
  \item{storewat}{Store drinking water (q10001)}
  \item{freechl}{Free chlorine detected in stored water (>0 mg/L)}
  \item{latseal}{Latrine has functional water seal (q809_9a)}
  \item{latfeces}{No visible feces on slab/floor of latrine}
  \item{humfeces}{No human feces observed in house/compound (q4201,4203,4205)}
  \item{hwsw}{Prim handwashing loc has water (q704_1)}
  \item{hwss}{Prim handwashing loc has soap (q704_2-6)}
  \item{hwsws}{Prim handwashing loc has water+soap (q704_1-6)}
  \item{lnsn}{LNS sachets consumed per week}
  \item{lnsp}{Percent of expected LNS sachets consumed}
  \item{rlnsn}{LNS sachets consumed per week (reported)}
  \item{rlnsp}{Percent of expected LNS sachets consumed (reported)}


  ...
}}
\source{
NA
}
\usage{
washb_bd_uptake
}
\description{
Bangladesh uptake analysis dataset (compound-svy obs), created by 6-bangladesh-d
}
\keyword{datasets}

