% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_glmFormat.R
\name{washb_glmFormat}
\alias{washb_glmFormat}
\title{washb_glmFormat}
\usage{
washb_glmFormat(fit, rfit, dmat, rowdropped, pair, vcovCL,
  family = "gaussian")
}
\arguments{
\item{fit}{glm fit object to be formatted}

\item{rfit}{output from coeftest with fit and sandwich SE}

\item{dmat}{dataframe used within the washb_glm function to fit data}

\item{rowdropped}{dummy vector indexing rows of data dropped because a var contained missing data, to be read through the function and added to the formatted output.}

\item{pair}{Pair-matched randomization ID variable (in WASH Benefits: block)}

\item{vcovCL}{sandwichSE function output}

\item{family}{GLM model family (gaussian, binomial, poisson, or negative binomial). Use "neg.binom" for Negative binomial.}
}
\value{
Returns a list of the risk ratios or risk differences, the variance-covariance matrix, and a vector indexing the rows of observations
        used to fit the glm model
}
\description{
Internal package function used to format the output of glm objects.
}
\examples{

}

