% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_lincom.R
\name{washb_lincom}
\alias{washb_lincom}
\title{washb_lincom}
\usage{
washb_lincom(lc, fit, vcv, measure = "RR")
}
\arguments{
\item{lc}{: Index vector of coefficients from glm modellinear combination of coefficients}

\item{fit}{: Model object returned from coeftest (class=coeftest) command within the washb_glm function. Accessed with $fit from washb_glm output.}

\item{vcv}{: variance-covariance matrix of coefficients. Accessed with $vcv from washb_glm output.}

\item{measure}{measure of effect. RR = risk ratio, RD = risk difference}
}
\value{
Returns a list of the risk ratios or risk differences, the variance-covariance matrix, and a vector indexing the rows of observations
        used to fit the glm model
}
\description{
Function to get Estimates, SEs, CIs, and P-values for Pr>|Z| from a linear combination of regression coefficients.
}
\examples{

}

