% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_mh.R
\name{washb_mh}
\alias{washb_mh}
\title{washb_mh}
\usage{
washb_mh(Y,tr,strat,contrast,measure="RR")

Estimate the Mantel-Haenszel prevalence ratio note: strata with no outcomes (i.e., missing PR) are dropped.
This is consistent with a fixed-effects regression analysis, in which those strata would not contribute to the estimates.
The arguments Y,tr,strat, below need to be from the same dataset:
}
\arguments{
\item{Y}{binary outcome variable (here: diar7d)}

\item{tr}{binary treatment group variable, comparison group first}

\item{strat}{stratification variable (here: block)}

\item{contrast}{vector of length 2 that includes the tr groups to contrast (control(reference arm) and then intervention)}

\item{measure}{measure of effect. RR = prev ratio, RD = prev difference}
}
\value{
to be written
}
\description{
Mantel-Haenszel Pooled estimates of the prevalence ratio (PR) or the prevalence
difference (PD) using randomization block as the stratification variable.
}
\details{
Function to call the M-H estimator for two different arms of the study this relies on
teh rma.mh() function in the metafor package.
}
\examples{
#Prescreen function applied to the Bangladesh diarrheal disease outcome.
#The function will test a matrix of covariates and return those related to child diarrheal disease with
#a <0.2 p-value from a likelihood ratio test.

Load cleaned diarrhea data (see vignette for cleaning steps):
library(washb)
data(washb_bd_diarClean)
ad<-washb_bd_diarClean

###Create vector of contrasts for each hypothesis to facilitate comparisons between arms.
#Hypothesis 1: Each intervention arm vs. Control
h1.contrasts <- list(
  c("Control","Water"),
  c("Control","Sanitation"),
  c("Control","Handwashing"),
  c("Control","WSH"),
  c("Control","Nutrition"),
  c("Control","Nutrition + WSH")
)

Apply washb_mh to the water vs. control arm contrast.
washb_mh(Y=ad$diar7d,tr=ad$tr, contrast=c("Control","Water"), strat=ad$block,measure="RR")

Return the risk difference instead of the risk ration:
washb_mh(Y=ad$diar7d,tr=ad$tr, contrast=c("Control","Water"), strat=ad$block,measure="RR")


Use sapply command to efficiently apply the function to all the treatment arm contrasts

diff.h1 <- t(sapply(h1.contrasts,washb_mh,Y=ad$diar7d,tr=ad$tr,strat=ad$block,measure="RR"))
rownames(diff.h1) <- c("Water v C","Sanitation v C","Handwashing v C","WSH v C","Nutrition v C","Nutrition + WSH v C")
print(diff.h1)

}

