% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_permute.R
\name{washb_permute}
\alias{washb_permute}
<<<<<<< HEAD
\title{WASH Benefits Wilcoxon Signed Rank permutation test function for two treatment arms conditional on randomization block.
washb_permute}
=======
\title{washb_permute}
>>>>>>> ad784f1a15d88f88351ea24f45f9ef11ca46bb47
\usage{
washb_permute(Y, tr, pair, contrast, nreps = 1e+05, seed = NULL)
}
\arguments{
\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{tr}{Binary treatment group variable (ideally a factor), comparison group first}

\item{pair}{Pair-matched randomization ID variable (in WASH Benefits: block)}

\item{contrast}{Vector of length 2 that includes the groups to contrast, e.g., c("Control","Water")}

\item{nreps}{Number of permutations to run.}

\item{seed}{Number for psuedo-random number generation in R}
}
<<<<<<< HEAD
\value{
to be written
}
\description{
WASH Benefits Wilcoxon Signed Rank permutation test function for two treatment arms conditional on randomization block.
washb_permute
}
\examples{
to be written
}
=======
\description{
washb_permute
}
>>>>>>> ad784f1a15d88f88351ea24f45f9ef11ca46bb47

