% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_ttest.R
\name{washb_ttest}
\alias{washb_ttest}
\title{washb_ttest}
\usage{
washb_ttest(Y,tr,strat,contrast)
}
\arguments{
\item{Y}{quantitative outcome variable (e.g. LAZ)}

\item{tr}{binary treatment group variable, comparison group first}

\item{strat}{stratification variable (here: block)}

\item{contrast}{vector of length 2 that includes the tr groups to contrast}
}
\value{

}
\description{
washb_ttest
}

