% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwichSE.R
\name{sandwichSE}
\alias{sandwichSE}
\title{sandwichSE}
\usage{
sandwichSE(dat, fm, cluster)
}
\arguments{
\item{dat}{Data used to fit the model}

\item{fm}{Model fit (object)}

\item{cluster}{Vector of cluster ID's}
}
\value{

}
\description{
sandwichSE
}
\examples{
 #to be written

}
\keyword{internal}
