% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_glm.R
\name{washb_glm}
\alias{washb_glm}
\title{Generalized linear model function for WASH Benefits study.
washb_glm}
\usage{
washb_glm(Y,tr,pair,W=NULL, forcedW=NULL, V=NULL, id,contrast,family="gaussian", pval=0.2, print=TRUE)
}
\arguments{
\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{tr}{Binary treatment group variable (ideally a factor), comparison group first}

\item{pair}{Pair-matched randomization ID variable (in WASH Benefits: block)}

\item{W}{Optional data frame that includes adjustment covariates (for adjusted estimates)}

\item{forcedW}{Optional vector of variable names to force as adjustment covariates (no screening)}

\item{V}{Optional variable name for subgroup analyses, which is interacted with 'tr'.}

\item{id}{ID variable for independent units (cluster ID)}

\item{contrast}{Vector of length 2 that includes the groups to contrast, e.g., c("Control","Water")}

\item{family}{GLM model family (gaussian, binomial, poisson, and negative binomial). Use "binonial(link='log')" to return prevalence ratios instead of odds ratios when the outcome is binary.  Use "neg.binom" for a Negative binomial model.}

\item{pval}{The p-value threshold: any variables with a p-value from the lielihood ratio test below this threshold will be returned. Defaults to 0.2}

\item{print}{Logical for whether to print function output, defaults to TRUE.}

\item{print}{Logical for whether to print names and descriptions of returned list objects}
}
\value{
Returns a list of the risk ratios or risk differences, the variance-covariance matrix, and a vector indexing the rows of observations
        used to fit the glm model
}
\description{
Note that in the WASH Benefits diarrheal disease primary outcome analysis, the glm model is fit with a log link,
 `family=binomial(link='log')`, to estimate prevalence ratios rather than with a logit link, `family="binomial"`,
 to estimate the odds ratio.
}
\details{
Occasionally, a glm model with a non-canonical link function like `family=binomial(link='log')` will fail to converge.
 If this occurs, use a modified poisson regression to estimate prevalence ratio using the argument `family=poisson(link='log')`.
 See [Zou 2004](http://www.uvm.edu/~rsingle/stat380/F04/possible/Zou-AJE-2004_PoissonRegBinaryData.pdf) for details.
}

