% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_glmFormat.R
\name{washb_glmFormat}
\alias{washb_glmFormat}
\title{washb_glmFormat}
\usage{
washb_glmFormat(rfit, RDfit = NULL, dmat, rowdropped, pair, vcovCL,
  vcovCL.rd = NULL, family = "gaussian", V = NULL, Subgroups = NULL,
  print = print)
}
\arguments{
\item{rfit}{output from coeftest with fit and sandwich SE}

\item{RDfit}{glm fit object with the identity link used to estimate risk differences.}

\item{dmat}{dataframe used within the washb_glm function to fit data}

\item{rowdropped}{dummy vector indexing rows of data dropped because a var contained missing data, to be read through the function and added to the formatted output.}

\item{pair}{Pair-matched randomization ID variable (in WASH Benefits: block)}

\item{vcovCL}{sandwichSE function output}

\item{vcovCL.rd}{sandwichSE function output for OLS RD model}

\item{family}{GLM model family (gaussian, binomial, poisson, or negative binomial). Use "neg.binom" for Negative binomial.}

\item{V}{Optional vector of variable names for subgroup analyses, which are interacted with 'tr'.}

\item{Subgroups}{Names of subgroups created by the interaction between treatment and V factor.}

\item{print}{Logical for whether to print function output}
}
\value{
Returns a list of the risk ratios or risk differences, the variance-covariance matrix, and a vector indexing the rows of observations
        used to fit the glm model
}
\description{
Internal package function used to format the output of glm objects.
}
\examples{

}

