% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_glmFormat.R
\name{washb_glmFormat}
\alias{washb_glmFormat}
\title{An internal formatting function for results output in glm objects}
\usage{
washb_glmFormat(glmModel = glmModel, rfit, dmat, rowdropped, contrast,
  pair, vcovCL, family = family, V = NULL, Subgroups = NULL,
  print = print, verbose = verbose)
}
\arguments{
\item{glmModel}{glm model fit passed through the function and added to the output list}

\item{rfit}{output from coeftest with fit and sandwich SE}

\item{dmat}{dataframe used within the washb_glm function to fit data}

\item{rowdropped}{dummy vector indexing rows of data dropped because a var contained missing data, to be read through the function and added to the formatted output.}

\item{contrast}{Vector of length 2 that includes the groups to contrast, e.g., c("Control","Water")}

\item{pair}{Pair-matched randomization ID variable (in WASH Benefits: block)}

\item{vcovCL}{sandwichSE function output}

\item{family}{GLM model family (gaussian, binomial, poisson, or negative binomial). Use "neg.binom" for Negative binomial.}

\item{V}{Optional vector of variable names for subgroup analyses, which are interacted with 'tr'.}

\item{Subgroups}{Names of subgroups created by the interaction between treatment and V factor.}

\item{print}{Logical for whether to print function output}

\item{verbose}{Logical for whether to print names and descriptions of returned list objects}
}
\value{
Returns a list of the risk ratios or risk differences, the variance-covariance matrix, and a vector indexing the rows of observations
        used to fit the glm model
}
\description{
Internal package function used to format the output of glm objects.
}
\examples{

}
\keyword{internal}
