% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_lincom.R
\name{washb_lincom}
\alias{washb_lincom}
\title{washb_lincom}
\usage{
washb_lincom(lc = NULL, varlist = NULL, fit, vcv, measure = "RR",
  flag = NULL)
}
\arguments{
\item{lc}{: Index vector of coefficients from glm modellinear combination of coefficients}

\item{varlist}{: Character vector of variables to include. Alternative to lc. If lc is specified, this arguement is ignored.}

\item{fit}{: Model object returned from coeftest (class=coeftest) command within the washb_glm function. Accessed with $fit from washb_glm output.}

\item{vcv}{: variance-covariance matrix of coefficients. Accessed with $vcv from washb_glm output.}

\item{measure}{measure of effect. RR = risk ratio, RD = risk difference}

\item{flag}{Internal argument used to flag and suppress printing if the washb_lincom function is called within another function.}
}
\value{
Returns a list of the risk ratios or risk differences, the variance-covariance matrix, and a vector indexing the rows of observations
        used to fit the glm model
}
\description{
Function to get Estimates, SEs, CIs, and P-values for Pr>|Z| from a linear combination of regression coefficients.
}
\examples{

}

