% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_mean.R
\name{washb_mean}
\alias{washb_mean}
\title{washb_mean}
\usage{
washb_mean(Y, id, print = FALSE)
}
\arguments{
\item{Y}{Outcome variable}

\item{id}{ID variable for independent units (in WASH Benefits: cluster ID)}

\item{print}{Logical. If equal to 'TRUE' it will print the results.}

\item{tr}{Variable that includes intervention group assignments (typically a factor)}

\item{group}{String. The arm within the tr variable within which the mean should be computed}
}
\value{
Returns a 1x4 matrix that includes the mean, robust SE, lower 95\% CI, upper 95\% CI
}
\description{
Calculate means for a variable along with robust sandwich SEs and 95\% confidence intervals that account for clustering within \texttt{id}
}

