% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_mh.R
\name{washb_mh}
\alias{washb_mh}
\title{washb_mh}
\usage{
washb_mh(Y,tr,strat,contrast,measure="RR")

Estimate the Mantel-Haenszel prevalence ratio note: strata with no outcomes (i.e., missing PR) are dropped.
This is consistent with a fixed-effects regression analysis, in which those strata would not contribute to the estimates.
The arguments Y,tr,strat, below need to be from the same dataset:
}
\arguments{
\item{Y}{binary outcome variable (here: diar7d)}

\item{tr}{binary treatment group variable, comparison group first}

\item{strat}{stratification variable (here: block)}

\item{contrast}{vector of length 2 that includes the tr groups to contrast (control(reference arm) and then intervention)}

\item{measure}{measure of effect. RR = prev ratio, RD = prev difference}
}
\value{
to be written
}
\description{
Mantel-Haenszel Pooled estimates of the prevalence ratio (PR) or the prevalence
difference (PD) using randomization block as the stratification variable.
}
\details{
Function to call the M-H estimator for two different arms of the study this relies on
teh rma.mh() function in the metafor package.
}

