% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_permute.R
\name{washb_permute}
\alias{washb_permute}
\title{washb_permute}
\usage{
washb_permute(Y, tr, pair, contrast, nreps = 1e+05, seed = NULL)
}
\arguments{
\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{tr}{Binary treatment group variable (ideally a factor), comparison group first}

\item{pair}{Pair-matched randomization ID variable (in WASH Benefits: block)}

\item{contrast}{Vector of length 2 that includes the groups to contrast, e.g., c("Control","Water")}

\item{nreps}{Number of permutations to run.}

\item{seed}{Number for psuedo-random number generation in R}
}
\description{
washb_permute
}

