% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_prescreen.R
\name{washb_prescreen}
\alias{washb_prescreen}
\title{washb_prescreen}
\usage{
washb_prescreen(Y, Ws, family = "gaussian", pval = 0.2, print = TRUE)
}
\arguments{
\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{Ws}{data frame that includes candidate adjustment covariates to screen}

\item{family}{GLM model family (gaussian, binomial, poisson, or negative binomial). Use "neg.binom" for Negative binomial.}

\item{pval}{The p-value threshold: any variables with a p-value from the lielihood ratio test below this threshold will be returned. Defaults to 0.2}

\item{print}{Logical for whether to print function output, defaults to TRUE.}
}
\value{
Function returns the list of variable names with a likelihood ratio test p-value <0.2 (unless a custom p-value is specified).
}
\description{
washb_prescreen
}
\examples{

#Prescreen function applied to the Bangladesh diarrheal disease outcome.
#The function will test a matrix of covariates and return those related to child diarrheal disease with
#a <0.2 p-value from a likelihood ratio test.

Load cleaned diarrhea data:
library(washb)
data(washb_bd_diarClean)


###Subset to a new dataframe the variables to be screened:
Ws <- subset(ad,select=c("fracode","month","agedays","sex","momage","momedu","momheight","hfiacat","Nlt18","Ncomp","watmin","elec","floor","walls","roof","asset_wardrobe","asset_table","asset_chair","asset_khat","asset_chouki","asset_tv","asset_refrig","asset_bike","asset_moto","asset_sewmach","asset_mobile"))

###Run the washb_prescreen function
prescreened_varnames<-washb_prescreen(Y=ad$diar7d,Ws,family="binomial")

###Rerun the function with a stricter p=value
prescreened_varname2s<-washb_prescreen(Y=ad$diar7d,Ws,family="binomial", pval=0.5)

}

