% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_prescreen.R
\name{washb_prescreen}
\alias{washb_prescreen}
\title{washb_prescreen}
\usage{
washb_prescreen(Y, Ws, family = "gaussian")
}
\arguments{
\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{Ws}{data frame that includes candidate adjustment covariates to screen}

\item{family}{GLM model family (gaussian, binomial, poisson, or negative binomial). Use "neg.binom" for Negative binomial.}
}
\description{
washb_prescreen
}

