% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_tmle.R
\name{washb_tmle}
\alias{washb_tmle}
\title{Target maximum likelihood estimation of intention-to-treat effects in the WASH Benefits trials}
\usage{
washb_tmle(Y, tr, W = NULL, id, Delta = rep(1, length(Y)),
  family = "gaussian", contrast, prtr = NULL, Q.SL.library = c("SL.mean",
  "SL.glm", "SL.bayesglm", "SL.gam", "SL.glmnet"),
  g.SL.library = Q.SL.library, pval = 0.2, seed = NULL)
}
\arguments{
\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{tr}{Binary treatment group variable, comparison group first}

\item{W}{Data frame that includes adjustment covariates}

\item{id}{ID variable for independent units. For pair-matched designs, this is the matched pair.}

\item{Delta}{indicator of missing outcome. 1 - observed, 0 - missing.}

\item{family}{Outcome family: \code{gaussian} (continuous outcomes, like LAZ) or \code{binomial} (binary outcomes like diarrhea or stunting)}

\item{contrast}{Vector of length 2 that includes the treatment groups to contrast (e.g., \code{contrast=c('Control','Nutrition')}).}

\item{prtr}{Vector of length 2 that includes the probability of receiving each treatment in the contrast argument for the main trial, the control is double sized (allocation 2:1) so control vs. active intervention should be \code{prtr=c(0.67,0.33)}. in contrasts where the arms are equally sized (allocation 1:1), then \code{prtr=c(0.5,0.5)}. If \code{prtr=NULL} (the default), then the treatment mechanism Pr(A|W) is estimated, which is usually more efficient (higher precision of the estimates).}

\item{g.SL.library}{Library of algorithms to include in the SuperLearner for the treatment model Pr(A|W) (ignored if prtr is specified), and for the missingness model Pr(Delta|A,W) (if Delta is specified)}

\item{pval}{The p-value threshold used to pre-screen covariates (\code{W}) based on a likelihood ratio test in a univariate regression with the outcome (\code{Y}). Variables with a univariate association p-value below this threshold will be used in the final model. Defaults to 0.2.}

\item{seed}{A seed for the pseudo-random cross-validation split used in model selection (use for perfectly reproducible results).}

\item{Q.SL.Library}{Library of algorithms to include in the SuperLearner for the outcome model}
}
\value{
A \code{tmle()} fit object (see the \code{\link[tmle]{tmle}} package for details). The \code{$estimates} list includes parameter estimates along with variance estimates and confidence intervals.
}
\description{
Estimate intention-to-treat parameters using targeted maximum likelihood estimation (TMLE), potentially adjusted for covariates and missing outcomes
}
\details{
The \code{washb_tmle} function is mainly a convenience wrapper for the \code{\link[tmle]{tmle}} function. It estimates intention-to-treat effects in a trial using targeted maximum likelihood estimation (TMLE). In brief, the function does the following: it restricts the data to complete observations in the two arms listed in the \code{contrast} argument, it pre-screens covariates (\code{W}), if specified, to select those that have a univariate association with the outcome, and then it estimates the intention-to-treat effect using TMLE.  If \code{family='binomial'}, then the function returns effects on the absolute, relative, and odds ratio scale. If \code{Delta} is specified (i.e., observations with missing outcomes are included), then the function will adjust the effects for missingness using inverse probability of censoring weights, with the weights estimated using super learning of \code{Pr(Delta|A,W)}.

If the analysis is pair-matched (as for primary outcomes), be sure to specify the pair ID in the \code{id} argument. Do not include pair IDs in the adjustment covariate set.

If adjustment covariates (\code{W}) are specified, then by default they are pre-screened and the subset that is associated with the outcome based on a likelihood ratio test are used in the estimation. There are some other important defaults to be aware of. First, if the argument \code{prtr=NULL} (the default), then the function estimates the treatment mechanism rather than treating it as known. Estimating the treatment mechanism theoretically increases precision and empirically we have found this as well (in the Bangladesh primary analysis). Second, the function uses the \code{\link[SuperLearner]{SuperLearner}} algorithm to adjust for covariates. The default algorithm library includes the simple mean, GLM, Bayes GLM with non-informative priors, generalized additive models (degree 2), and lasso (glmnet). You can type \code{listWrappers()} to see the full set of algorithms implemented in the super learner. If you just wish to use a main effects GLM model to adjust for the covariates, then you can specify \code{Q.SL.library="SL.glm"}.

If you want to adjust for missing outcomes in the analysis, then you need to include observations that have a missing outcome (\code{Y}) with \code{Delta=0} for those observations. Observations with missing outcomes should have treatment (\code{tr}) and covariate (\code{W}) information, which are used to create weights for \code{Pr(Delta|A,W)}.

Note: this function depends on the \code{\link[tmle]{tmle}} package, the \code{\link[SuperLearner]{SuperLearner}} package, as well as the internal washb_prescreen() and design_matrix() functions.
}
\examples{
# TBD

}
\references{
Gruber S, van der Laan M. tmle: An R Package for Targeted Maximum Likelihood Estimation. J Stat Softw. 2012;51: 1–35. \link{https://www.jstatsoft.org/article/view/v051i13}

Balzer LB, van der Laan MJ, Petersen ML, SEARCH Collaboration. Adaptive pre-specification in randomized trials with and without pair-matching. Stat Med. 2016; doi:10.1002/sim.7023 \link{http://onlinelibrary.wiley.com/doi/10.1002/sim.7023/abstract}
}

