% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_tmle.R
\name{washb_tmle}
\alias{washb_tmle}
\title{Target maximum likelihood estimation function for WASH Benefits study.
washb_tmle}
\usage{
washb_tmle(Y, tr, strat, W, id, contrast, prtr = c(0.5, 0.5),
  modelfit = "glm", family = "gaussian", SL.library = c("SL.mean",
  "SL.glm", "SL.bayesglm", "SL.gam", "SL.glmnet"), seed = NULL)
}
\arguments{
\item{Y}{Outcome variable (continuous, such as LAZ, or binary, such as diarrhea)}

\item{tr}{Binary treatment group variable, comparison group first}

\item{strat}{Stratification variable (here: block)}

\item{W}{Data frame that includes adjustment covariates}

\item{id}{ID variable for independent units (e.g., cluster)}

\item{contrast}{Vector of length 2 that includes the tr groups to contrast}

\item{prtr}{Vector of length 2 that includes the probability of receiving each treatment in the contrast argument for the main trial, the control is double sized (allocation 2:1) so control vs. active intervention should be prtr=c(0.67,0.33). in contrasts where the arms are equally sized (allocation 1:1), then prtr=c(0.5,0.5).}

\item{modelfit}{String variable of value "glm" for a glm prediction of E(Y|A,W) and E(A|W), or "sl" for a SuperLearner prediction.}

\item{family}{Outcome family: gaussian (continuous outcomes, like LAZ) or binomial (binary outcomes like diarrhea or stunting)}

\item{seed}{A seed for the pseudo-random CV split for perfectly reproducible results}

\item{SL.Library}{Library of algorithms to include in the SuperLearner (pre-specified defaults are encoded above)}
}
\value{
To be filled
}
\description{
# Targeted maximum likelihood estimation (TMLE) for the WASH Benefits intention to treat (ITT) analyses
}
\details{
The function does the following:
}
\examples{


}

