% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/washb_ttest.R
\name{washb_ttest}
\alias{washb_ttest}
\title{washb_ttest}
\usage{
washb_ttest(Y,tr,strat,contrast)
}
\arguments{
\item{Y}{quantitative outcome variable (e.g. LAZ)}

\item{tr}{binary treatment group variable, comparison group first}

\item{strat}{stratification variable (here: block)}

\item{contrast}{vector of length 2 that includes the tr groups to contrast}
}
\value{
data frame with mean difference ("diff"), 95% confidence intervals ("ci.lb" and "ci.ub"), t-statistic ("t-stat"), and p-value("p")
}
\description{
Function to call the paired t-test for two different arms of the study.
Estimates the paired t-test for differences in means paired within randomization blocks.
The arguments Y,tr,strat, below need to be from the same dataset:
}
\examples{
 #The washb_ttest function

 #Load in Bangladesh cleaned anthropometry data. (See vignette for specific cleaning steps).
 data(washb_bd_anthroClean)
 ad<-washb_bd_anthroClean

 #Run paired ttest function for water vs. control comparison:
 washb_ttest(Y=ad$laz,tr=ad$tr,strat=ad$block, contrast=c("Control","Water"))
}

